Refactor: 전시리뷰목록 컴포넌트 폴더 분리, 파일명 수정
