+++ src/components/mypage/WrittenBlogReview.tsx
@@ -5,7 +5,7 @@ import { useQuery } from "react-query";
 import { BlogReviewListRes, myBlogReviewsApi } from "../../apis/member";
 import useGetNewTokenApi from "../../apis/useGetRefreshToken";
 import isApiError from "../../utils/isApiError";
-import BlogReviewCard from "../blogreview/presentation/BlogReviewCard";
+import BlogReviewCard from "../blogReviewList/ReviewCard";
 import Pagination from "../Pagination";
 
 const WrittenBlogReview = () => {
